+++ CHANGES.md
@@ -1,5 +1,7 @@
 # Changes
 
+### cordova-sqlite-storage-commoncore 1.0.0-pre1
+
 ### cordova-sqlite-storage 2.5.1
 
 - fix internal plugin cleanup error log on Android