+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Changes
 
-#### cordova-sqlite-storage 3.2.2-dev
+#### cordova-sqlite-storage 3.3.0
 
 - new default page & cache sizes with cordova-sqlite-storage-dependencies@2.1.0